requirements.txt
requests
pydantic
python-dotenv
pytest
pyyaml
prometheus-client

